Revert "Improve IntroComp"
