+++ react-vite/package.json
@@ -10,7 +10,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "framer-motion": "^12.0.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
