--- react-vite/src/components/intro/introduce-section.css
@@ -1,44 +1,0 @@
-/* 写真ボックス */
-.image-box {
-    flex-shrink: 0;
-    width: 300px;
-    height: 300px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 2px solid #ddd;
-    border-radius: 50%;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    margin: 0 60px;
-}
-
-.image-box img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 50%;
-}
-
-/* 文章ボックス */
-.text-box {
-    flex-direction: column;
-    justify-content: center;
-    padding: 16px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    background-color: rgba(224, 224, 224, 0.479);
-}
-
-.text-box h2 {
-    margin: 0;
-    padding-bottom: 15px;
-    font-size: 35px;
-    color: #696969;
-}
-
-.text-box p {
-    margin: 0;
-    font-size: 18px;
-    color: #696969;
-}
