+++ react-vite/src/components/top/top.css
@@ -27,7 +27,7 @@
   .top-container h1 {
         font-family: 'Arial', sans-serif; /* 好きなフォントに変更可 */
         font-size: 2.5rem; /* 適宜調整 */
-        color: #e6e6e6; /* 白文字 */
+        color: white; /* 白文字 */
         text-align: center; /* 中央揃え */
         text-shadow: 0 2px 5px rgba(0, 0, 0, 0.5); /* 影を追加 */
         padding: 10px;
